Add mock API and social login simulation
